Add functionality to automatically purge points that go over the limit.
https://github.com/fermi-ad/plotting-app/issues/95

+++ lib/entities/plot_metadata.dart
@@ -9,7 +9,7 @@ class PlotMetadata with ChangeNotifier {
   double? _latestRequestEpochTime;
   double? _latestDataEpochTime;
 
-  double plotDataBytes = 0;
+  int plotDataBytes = 0;
 
   PlotMetadata();
